fix: moved switchNetwok to new hook
